refactor: fold base 3D view into volume viewer
